# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Frontline Prep API",
    "description": "Backend for the ISSB preparation platform: catalog, question bank, mock tests, and the ISSB simulation suite. Question generation runs locally -- no external AI service is called.",
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/v1/questions": {
      "get": {
        "tags": [
          "practice"
        ],
        "summary": "Browse Questions",
        "description": "Approved questions only, and never with the answer key attached.",
        "operationId": "browse_questions_api_v1_questions_get",
        "parameters": [
          {
            "name": "module_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Module Id"
            }
          },
          {
            "name": "topic_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Topic Id"
            }
          },
          {
            "name": "difficulty",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Difficulty"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Difficulty"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 10000,
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_QuestionOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/questions/{question_id}/report": {
      "post": {
        "tags": [
          "practice"
        ],
        "summary": "Report Question",
        "operationId": "report_question_api_v1_questions__question_id__report_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "question_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Question Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuestionReportIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionReportOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tests": {
      "get": {
        "tags": [
          "practice"
        ],
        "summary": "List Tests",
        "operationId": "list_tests_api_v1_tests_get",
        "parameters": [
          {
            "name": "service_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Service Id"
            }
          },
          {
            "name": "is_mock",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Is Mock"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TestTemplateOut"
                  },
                  "title": "Response List Tests Api V1 Tests Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/attempts": {
      "post": {
        "tags": [
          "practice"
        ],
        "summary": "Start Attempt",
        "operationId": "start_attempt_api_v1_attempts_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AttemptStartIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttemptOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "practice"
        ],
        "summary": "Attempt History",
        "operationId": "attempt_history_api_v1_attempts_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 10000,
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_AttemptSummaryOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/attempts/{attempt_id}": {
      "get": {
        "tags": [
          "practice"
        ],
        "summary": "Resume Attempt",
        "operationId": "resume_attempt_api_v1_attempts__attempt_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "attempt_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Attempt Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttemptOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/attempts/{attempt_id}/submit": {
      "post": {
        "tags": [
          "practice"
        ],
        "summary": "Submit Attempt",
        "operationId": "submit_attempt_api_v1_attempts__attempt_id__submit_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "attempt_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Attempt Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AttemptSubmitIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttemptResultOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/attempts/{attempt_id}/result": {
      "get": {
        "tags": [
          "practice"
        ],
        "summary": "Attempt Result",
        "operationId": "attempt_result_api_v1_attempts__attempt_id__result_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "attempt_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Attempt Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttemptResultOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/attempts/{attempt_id}/abandon": {
      "post": {
        "tags": [
          "practice"
        ],
        "summary": "Abandon Attempt",
        "operationId": "abandon_attempt_api_v1_attempts__attempt_id__abandon_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "attempt_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Attempt Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Msg"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AttemptOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "mode": {
            "type": "string",
            "title": "Mode"
          },
          "status": {
            "$ref": "#/components/schemas/AttemptStatus"
          },
          "started_at": {
            "type": "string",
            "format": "date-time",
            "title": "Started At"
          },
          "expires_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "total_questions": {
            "type": "integer",
            "title": "Total Questions"
          },
          "duration_min": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Min"
          },
          "template": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TestTemplateOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "sections": {
            "items": {
              "$ref": "#/components/schemas/AttemptSectionOut"
            },
            "type": "array",
            "title": "Sections",
            "default": []
          },
          "questions": {
            "items": {
              "$ref": "#/components/schemas/QuestionOut"
            },
            "type": "array",
            "title": "Questions",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "id",
          "mode",
          "status",
          "started_at",
          "total_questions"
        ],
        "title": "AttemptOut",
        "description": "The live paper. Carries questions without answer keys."
      },
...
      "AttemptResultOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "status": {
            "$ref": "#/components/schemas/AttemptStatus"
          },
          "mode": {
            "type": "string",
            "title": "Mode"
          },
          "submitted_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Submitted At"
          },
          "duration_sec": {
            "type": "integer",
            "title": "Duration Sec"
          },
          "total_questions": {
            "type": "integer",
            "title": "Total Questions"
          },
          "attempted": {
            "type": "integer",
            "title": "Attempted"
          },
          "correct": {
            "type": "integer",
            "title": "Correct"
          },
          "wrong": {
            "type": "integer",
            "title": "Wrong"
          },
          "score": {
            "type": "number",
            "title": "Score"
          },
          "max_score": {
            "type": "number",
            "title": "Max Score"
          },
          "percentage": {
            "type": "number",
            "title": "Percentage"
          },
          "passed": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Passed"
          },
          "topic_breakdown": {
            "additionalProperties": true,
            "type": "object",
            "title": "Topic Breakdown",
            "default": {}
          },
          "weak_topics": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Weak Topics",
            "default": []
          },
          "review": {
            "items": {
              "$ref": "#/components/schemas/AttemptQuestionResult"
            },
            "type": "array",
            "title": "Review",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "id",
          "status",
          "mode",
          "duration_sec",
          "total_questions",
          "attempted",
          "correct",
          "wrong",
          "score",
          "max_score",
          "percentage"
        ],
        "title": "AttemptResultOut"
      },
...
      "AttemptStartIn": {
        "properties": {
          "template_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Template Id"
          },
          "module_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Module Id"
          },
          "topic_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Topic Id"
          },
          "count": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Count",
            "default": 20
          },
          "difficulty": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Difficulty"
              },
              {
                "type": "null"
              }
            ]
          },
          "mode": {
            "type": "string",
            "maxLength": 16,
            "title": "Mode",
            "default": "practice"
          },
          "only_weak": {
            "type": "boolean",
            "title": "Only Weak",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "AttemptStartIn",
        "description": "Start from a template, or from a module for an ad-hoc drill."
      },
      "AttemptStatus": {
        "type": "string",
        "enum": [
          "in_progress",
          "submitted",
          "expired",
          "abandoned"
        ],
        "title": "AttemptStatus"
      },
      "AttemptSubmitIn": {
        "properties": {
          "answers": {
            "items": {
              "$ref": "#/components/schemas/AnswerIn"
            },
            "type": "array",
            "maxItems": 300,
            "title": "Answers"
          },
          "duration_sec": {
            "type": "integer",
            "maximum": 86400.0,
            "minimum": 0.0,
            "title": "Duration Sec",
            "default": 0
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "AttemptSubmitIn"
      },
...
      "Difficulty": {
        "type": "string",
        "enum": [
          "easy",
          "medium",
          "hard"
        ],
        "title": "Difficulty"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Msg": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "Msg"
      },
...
      "Page_AttemptSummaryOut_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AttemptSummaryOut"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "size": {
            "type": "integer",
            "title": "Size"
          },
          "pages": {
            "type": "integer",
            "title": "Pages"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "size",
          "pages"
        ],
        "title": "Page[AttemptSummaryOut]"
      },
...
      "Page_QuestionOut_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/QuestionOut"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "size": {
            "type": "integer",
            "title": "Size"
          },
          "pages": {
            "type": "integer",
            "title": "Pages"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "size",
          "pages"
        ],
        "title": "Page[QuestionOut]"
      },
...
      "QuestionReportIn": {
        "properties": {
          "reason": {
            "type": "string",
            "maxLength": 40,
            "title": "Reason"
          },
          "note": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 400
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "reason"
        ],
        "title": "QuestionReportIn"
      },
      "QuestionReportOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "question_id": {
            "type": "integer",
            "title": "Question Id"
          },
          "user_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Id"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "resolved": {
            "type": "boolean",
            "title": "Resolved"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "question_id",
          "reason",
          "resolved",
          "created_at"
        ],
        "title": "QuestionReportOut"
      },
...
      "TestTemplateOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "service_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Service Id"
          },
          "stage_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stage Id"
          },
          "program_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Program Id"
          },
          "sections": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Sections",
            "default": []
          },
          "duration_min": {
            "type": "integer",
            "title": "Duration Min"
          },
          "total_questions": {
            "type": "integer",
            "title": "Total Questions"
          },
          "pass_percentage": {
            "type": "number",
            "title": "Pass Percentage"
          },
          "negative_marking": {
            "type": "number",
            "title": "Negative Marking"
          },
          "show_answers_after": {
            "type": "boolean",
            "title": "Show Answers After"
          },
          "is_mock": {
            "type": "boolean",
            "title": "Is Mock"
          },
          "is_free": {
            "type": "boolean",
            "title": "Is Free"
          },
          "sort_order": {
            "type": "integer",
            "title": "Sort Order",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "id",
          "slug",
          "title",
          "duration_min",
          "total_questions",
          "pass_percentage",
          "negative_marking",
          "show_answers_after",
          "is_mock",
          "is_free"
        ],
        "title": "TestTemplateOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "description": "JWT access token",
        "scheme": "bearer"
      }
    }
  }
}